+++ packages/components/modal/src/Modal.styles.ts
@@ -28,7 +28,7 @@ export const Description = styled(Typography.Text)`
   font-weight: normal;
   display: block;
   padding: 12px 0 0;
-  margin: 8px 0 -14px;
+  margin: 8px 0 0;
 
   background-image: linear-gradient(
     to right,