+++ packages/components/tooltip/src/Tooltip.types.ts
@@ -1,18 +1,12 @@
 import * as React from 'react';
 
 export type tooltipTypes = 'default' | 'icon' | 'largeSimple' | 'tutorial';
-
+export type descriptionType = string /* | [string] */ | React.ReactNode /* | [React.ReactNode] */;
+// TODO support descriptionType as array of element
 export default interface TooltipExtendedProps {
   type?: tooltipTypes;
   icon?: React.ReactNode;
-  title?: string;
-  description?: string;
+  title?: string | React.ReactNode;
+  description?: descriptionType;
   theme: { [k: string]: string };
 }
-
-export const typesOverlayClasses = {
-  default: 'defaultOverlayClass',
-  icon: 'iconOverlayClass',
-  largeSimple: 'largeSimpleOverlayClass',
-  tutorial: 'tutorialOverlayClass',
-};