+++ packages/components/block/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-block@0.1.4...@synerise/ds-block@0.1.5) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-block
+
+## [0.1.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-block@0.1.3...@synerise/ds-block@0.1.4) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-block
+
+## [0.1.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-block@0.1.2...@synerise/ds-block@0.1.3) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-block
+
+## [0.1.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-block@0.1.1...@synerise/ds-block@0.1.2) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-block
+
 ## [0.1.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-block@0.1.0...@synerise/ds-block@0.1.1) (2019-12-10)
 
 **Note:** Version bump only for package @synerise/ds-block