+++ packages/components/card-select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.21](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-select@0.1.20...@synerise/ds-card-select@0.1.21) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-card-select
+
+## [0.1.20](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-select@0.1.19...@synerise/ds-card-select@0.1.20) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-card-select
+
+## [0.1.19](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-select@0.1.18...@synerise/ds-card-select@0.1.19) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-card-select
+
+## [0.1.18](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-select@0.1.17...@synerise/ds-card-select@0.1.18) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-card-select
+
 ## [0.1.17](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-select@0.1.16...@synerise/ds-card-select@0.1.17) (2019-12-10)
 
 **Note:** Version bump only for package @synerise/ds-card-select