+++ packages/components/file-uploader/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-file-uploader@0.2.6...@synerise/ds-file-uploader@0.2.7) (2020-01-08)
+
+**Note:** Version bump only for package @synerise/ds-file-uploader
+
+## [0.2.6](https://github.com/Synerise/synerise-design/compare/@synerise/ds-file-uploader@0.2.5...@synerise/ds-file-uploader@0.2.6) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-file-uploader
+
+## [0.2.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-file-uploader@0.2.4...@synerise/ds-file-uploader@0.2.5) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-file-uploader
+
+## [0.2.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-file-uploader@0.2.3...@synerise/ds-file-uploader@0.2.4) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-file-uploader
+
+## [0.2.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-file-uploader@0.2.2...@synerise/ds-file-uploader@0.2.3) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-file-uploader
+
 ## [0.2.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-file-uploader@0.2.1...@synerise/ds-file-uploader@0.2.2) (2019-12-17)
 
 **Note:** Version bump only for package @synerise/ds-file-uploader