+++ packages/components/icon/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.7.0...@synerise/ds-icon@0.7.1) (2020-01-02)
+
+### Bug Fixes
+
+- **icon:** add dynamic loading for icons in storybook ([fc277b8](https://github.com/Synerise/synerise-design/commit/fc277b824728b200c979ecf234eb1f137be253e2))
+
+# [0.7.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.6.2...@synerise/ds-icon@0.7.0) (2019-12-30)
+
+### Features
+
+- **icon:** add custom block icons ([7fbbdc8](https://github.com/Synerise/synerise-design/commit/7fbbdc815ab6dfcb83accbbf21b0291a00568121))
+
+## [0.6.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.6.1...@synerise/ds-icon@0.6.2) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-icon
+
+## [0.6.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.6.0...@synerise/ds-icon@0.6.1) (2019-12-18)
+
+### Bug Fixes
+
+- 🐛 Added missing icons, storybook update script ([e37101d](https://github.com/Synerise/synerise-design/commit/e37101d4eddfa56c5c5b8e5bd8db5b4f0f42bed8))
+
 # [0.6.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.5.0...@synerise/ds-icon@0.6.0) (2019-12-10)
 
 ### Features