+++ packages/components/manageable-list/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.3.4...@synerise/ds-manageable-list@0.3.5) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-manageable-list
+
+## [0.3.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.3.3...@synerise/ds-manageable-list@0.3.4) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-manageable-list
+
+## [0.3.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.3.2...@synerise/ds-manageable-list@0.3.3) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-manageable-list
+
+## [0.3.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.3.1...@synerise/ds-manageable-list@0.3.2) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-manageable-list
+
 ## [0.3.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.3.0...@synerise/ds-manageable-list@0.3.1) (2019-12-10)
 
 **Note:** Version bump only for package @synerise/ds-manageable-list