+++ packages/components/message/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-message@0.2.0...@synerise/ds-message@0.2.1) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-message
+
 # [0.2.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-message@0.1.1...@synerise/ds-message@0.2.0) (2019-12-06)
 
 ### Features