+++ packages/components/page-header/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.7](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.6.6...@synerise/ds-page-header@0.6.7) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+## [0.6.6](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.6.5...@synerise/ds-page-header@0.6.6) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+## [0.6.5](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.6.4...@synerise/ds-page-header@0.6.5) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+## [0.6.4](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.6.3...@synerise/ds-page-header@0.6.4) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
 ## [0.6.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.6.2...@synerise/ds-page-header@0.6.3) (2019-12-17)
 
 **Note:** Version bump only for package @synerise/ds-page-header