+++ packages/components/pagination/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.5](https://github.com/synerise/synerise-design/compare/@synerise/ds-pagination@0.5.4...@synerise/ds-pagination@0.5.5) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-pagination
+
+## [0.5.4](https://github.com/synerise/synerise-design/compare/@synerise/ds-pagination@0.5.3...@synerise/ds-pagination@0.5.4) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-pagination
+
+## [0.5.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-pagination@0.5.2...@synerise/ds-pagination@0.5.3) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-pagination
+
+## [0.5.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-pagination@0.5.1...@synerise/ds-pagination@0.5.2) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-pagination
+
 ## [0.5.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-pagination@0.5.0...@synerise/ds-pagination@0.5.1) (2019-12-10)
 
 **Note:** Version bump only for package @synerise/ds-pagination