+++ packages/components/slider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.4...@synerise/ds-slider@0.2.5) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+## [0.2.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.3...@synerise/ds-slider@0.2.4) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+## [0.2.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.2...@synerise/ds-slider@0.2.3) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+## [0.2.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.1...@synerise/ds-slider@0.2.2) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
 ## [0.2.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.2.0...@synerise/ds-slider@0.2.1) (2019-12-17)
 
 **Note:** Version bump only for package @synerise/ds-slider