+++ packages/components/status/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.26](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.25...@synerise/ds-status@0.0.26) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.25](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.24...@synerise/ds-status@0.0.25) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.24](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.23...@synerise/ds-status@0.0.24) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.23](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.22...@synerise/ds-status@0.0.23) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-status
+
 ## [0.0.22](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.21...@synerise/ds-status@0.0.22) (2019-12-10)
 
 **Note:** Version bump only for package @synerise/ds-status