+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-switch@0.1.0...@synerise/ds-switch@0.1.1) (2019-12-23)
+
+### Bug Fixes
+
+- export props from switch & input number components ([4e7546c](https://github.com/synerise/synerise-design/commit/4e7546c93748c8e8db02bf5ed4676c253729e244))
+
 # [0.1.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-switch@0.0.15...@synerise/ds-switch@0.1.0) (2019-12-06)
 
 ### Features