+++ packages/components/tabs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.5](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.6.4...@synerise/ds-tabs@0.6.5) (2020-01-02)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
+## [0.6.4](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.6.3...@synerise/ds-tabs@0.6.4) (2019-12-30)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
+## [0.6.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.6.2...@synerise/ds-tabs@0.6.3) (2019-12-23)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
+## [0.6.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.6.1...@synerise/ds-tabs@0.6.2) (2019-12-18)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
 ## [0.6.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.6.0...@synerise/ds-tabs@0.6.1) (2019-12-10)
 
 **Note:** Version bump only for package @synerise/ds-tabs