+++ packages/docs-site/website/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://gitlab.synerise.com/Frontend/synerise-design/compare/docs@0.1.0...docs@0.1.1) (2019-12-23)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
 # [0.1.0](https://gitlab.synerise.com/Frontend/synerise-design/compare/docs@0.0.11...docs@0.1.0) (2019-12-06)